Move ResourceType to data.py and create_client to client.py
